bot/requirements.txt
python-telegram-bot>=21.0.1,<22.0.0
